+++ CHANGELOG.md
@@ -15,9 +15,16 @@ different versioning scheme, following the Haskell community's
 * `gbc` & compiler library: TBD
 * IDL core version: TBD
 * IDL comm version: TBD
-* C++ version: TBD
+* C++ version: TBD (bug fix bump needed)
 * C# NuGet version: TBD  (bug fix bump needed)
 * C# Comm NuGet version: TBD
+
+### C++ ###
+
+* When Unicode conversion fails during JSON deserialization to wstring, throw a 
+  Bond CoreException instead of a Boost exception.
+* Throw Bond CoreException when SimpleJSON deserializes map key with no matching value.
+* Throw Bond CoreException when SimpleJSON deserializes map key of non-primitive type.
 
 ### C# ###